backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic
numpy
scikit-learn
pypdf
python-docx
python-multipart
sentence-transformers
keybert
